Beta 12

+++ database/migrations/2021_07_30_122338_create_pago_pedidos_table.php
@@ -23,7 +23,7 @@ class CreatePagoPedidosTable extends Migration
                // 4 Credito  
                // 5 Otros
                // 6 vuelto
-            $table->float("monto",10,2);
+            $table->decimal("monto",8,2);
             $table->boolean('cuenta')->default(1);
             // 1 es credito 
             // 0 es abono
